Refactor home component and create uploadVideo component
