+++ frontend/src/components/upload_video/upload_video_container.js
@@ -1,0 +1,14 @@
+import { connect } from "react-redux";
+
+import UploadVideo from "./upload_video";
+
+const mSP = ({ session }) => {
+  return {
+    currentUser: session
+  };
+};
+
+export default connect(
+  mSP,
+  null
+)(UploadVideo);
